newPage changed to Form
