--- src/components/NewGame.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const NewGame = (props) => {
-    return(
-        <div>
-            <h2>Let's start a new game!</h2>
-        </div>
-    )
-}
-
-export default NewGame;
